sprint_2: feat: разбил страницу чатов на два компонента
