+++ src/components/DialogList/DialogList.module.pcss
@@ -1,0 +1,33 @@
+.panel-chat {
+    display: flex;
+    flex-direction: column;
+    max-width: 310px;
+    width: 100%;
+    background-color: #fbfbfb;
+    border-right: 1px solid #eaeaea;
+}
+
+.sub-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding-left: 7px;
+}
+
+.panel-header {
+    display: flex;
+    margin: 21px 10px 14px;
+    flex-direction: column;
+}
+
+.search-input {
+    margin-top: 23px;
+}
+
+.list-chat {
+    margin: 0 10px;
+    list-style-type: none;
+    padding: 0;
+}
+
+
