+++ src/components/MessageBlock/MessageBlock.module.pcss
@@ -1,0 +1,61 @@
+.chat {
+    display: flex;
+    flex-direction: column;
+    margin: 10px 20px 11px;
+    width: calc(100% - 40px);
+    height: calc(100% - 22px);
+}
+
+.chat-header {
+    display: flex;
+    align-items: center;
+    margin-bottom: 10px;
+}
+
+.chat-img {
+    width: 34px;
+    height: 34px;
+}
+
+.icon-menu {
+    margin-left: auto;
+    margin-right: 20px;
+}
+
+.chat-name {
+    margin-left: 10px;
+    font-size: 13px;
+    line-height: 13px;
+    color: #1e1e1e;
+}
+
+.correspondence {
+    display: flex;
+    justify-content: flex-end;
+    flex-direction: column;
+    height: 100%;
+    width: 100%;
+    border-top: 1px solid #eaeaea;
+    border-bottom: 1px solid #eaeaea;
+}
+
+.send-msg-form {
+    display: flex;
+    align-items: center;
+}
+
+.send-msg-input {
+    display: block;
+    margin: 11px 10px;
+    padding: 10px 20px;
+    border-radius: 30px;
+    width: 100%;
+    background-color: #efefef;
+    color: var(--grey-color);
+    border: none;
+    outline: none;
+}
+
+.send-msg-submit {
+    transform: rotate(180deg);
+}
