+++ src/components/Modal/Modal.module.pcss
@@ -4,6 +4,7 @@
     left: 0;
     bottom: 0;
     right: 0;
+    z-index: 1;
     display: flex;
     justify-content: center;
     align-items: center;
