+++ src/page/ChatPage/ChatPage.module.pcss
@@ -3,91 +3,3 @@
     height: 100vh;
     width: 100vw;
 }
-
-.left-panel-chat {
-    display: flex;
-    flex-direction: column;
-    max-width: 310px;
-    width: 100%;
-    background-color: #fbfbfb;
-    border-right: 1px solid #eaeaea;
-}
-
-.left-panel-header {
-    display: flex;
-    margin: 21px 10px 14px;
-    flex-direction: column;
-}
-
-.search-input {
-    margin-top: 23px;
-}
-
-.list-chat {
-    margin: 0 10px;
-    list-style-type: none;
-    padding: 0;
-}
-
-.chat {
-    display: flex;
-    flex-direction: column;
-    margin: 10px 20px 11px;
-    width: calc(100% - 40px);
-    height: calc(100% - 22px);
-}
-
-.chat-header {
-    display: flex;
-    align-items: center;
-    margin-bottom: 10px;
-}
-
-.chat-img {
-    width: 34px;
-    height: 34px;
-}
-
-.icon-menu {
-    margin-left: auto;
-    margin-right: 20px;
-}
-
-.chat-name {
-    margin-left: 10px;
-    font-size: 13px;
-    line-height: 13px;
-    color: #1e1e1e;
-}
-
-.correspondence {
-    display: flex;
-    justify-content: flex-end;
-    flex-direction: column;
-    height: 100%;
-    width: 100%;
-    border-top: 1px solid #eaeaea;
-    border-bottom: 1px solid #eaeaea;
-}
-
-.send-msg-form {
-    display: flex;
-    align-items: center;
-}
-
-.send-msg-input {
-    display: block;
-    margin: 11px 10px;
-    padding: 10px 20px;
-    border-radius: 30px;
-    width: 100%;
-    background-color: #efefef;
-    color: var(--grey-color);
-    border: none;
-    outline: none;
-}
-
-.send-msg-submit {
-    transform: rotate(180deg);
-}
-
